Moved go to its own module and changed start selection
A new subdirectory "modules" has been created. It
will keep all the files to the modules that will be
installed. By now, go language installation has been
moved to it.

Now at the very beginning this directory is scanned
for the known modules and the files are loaded and
registered in the MODULES var.

When a module name is passed trough the command line,
if it is known, the function install_${MODULE_NAME} is
executed. This will be the entry point for the module
installation.

The name of the file in the "modules" subdir must match
with the name of the command line option. E.g. to install
golang, you will call "install.sh golang" and a module
file must exist in "modules" named "golang.sh".